+++ AutoFieldTranslationExperiment.sln.DotSettings.user
@@ -2,16 +2,14 @@
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=7020124F_002D9FFC_002D4AC3_002D8F3D_002DAAB8E0240759_002Ff_003AConfigurationExtensions_002Ecs_002Fl_003A_002E_002E_003F_002E_002E_003F_002E_002E_003FLibrary_003FApplication_0020Support_003FJetBrains_003FRider2024_002E2_003Fresharper_002Dhost_003FSourcesCache_003Fb8c1c57742337e59591581af3eb2c647e9ed52d2951655fd74b1facf3ff520_003FConfigurationExtensions_002Ecs/@EntryIndexedValue">ForceIncluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=7020124F_002D9FFC_002D4AC3_002D8F3D_002DAAB8E0240759_002Ff_003AResults_002Ecs_002Fl_003A_002E_002E_003F_002E_002E_003F_002E_002E_003FLibrary_003FApplication_0020Support_003FJetBrains_003FRider2024_002E2_003Fresharper_002Dhost_003FSourcesCache_003Fa434b74a44cbf5195ce762afb1f966a5e1ce8e28090ea5b507b188fa8d35_003FResults_002Ecs/@EntryIndexedValue">ForceIncluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=7020124F_002D9FFC_002D4AC3_002D8F3D_002DAAB8E0240759_002Ff_003AServiceProvider_002Ecs_002Fl_003A_002E_002E_003F_002E_002E_003F_002E_002E_003FLibrary_003FApplication_0020Support_003FJetBrains_003FRider2024_002E2_003Fresharper_002Dhost_003FSourcesCache_003F7d81b2d4f22bee75e5438c707251ae43cb0974c207db91ffc159118c84b4eb9_003FServiceProvider_002Ecs/@EntryIndexedValue">ForceIncluded</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=91636bcb_002D6b9a_002D420f_002D999f_002Dfe0b1d09b214/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="Returns_AllProducts" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;Project Location="/Users/dominicodonnell/Dev/PERSONAL/AutoFieldTranslationExperiment/test/Web.Integration" Presentation="&amp;lt;test&amp;gt;\&amp;lt;Web.Integration&amp;gt;" /&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=91636bcb_002D6b9a_002D420f_002D999f_002Dfe0b1d09b214/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="Returns_AllProducts" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;Project Location="\Users\dominicodonnell\Dev\PERSONAL\AutoFieldTranslationExperiment\test\Web.Integration" Presentation="&amp;lt;test&amp;gt;\&amp;lt;Web.Integration&amp;gt;" /&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=94ce6ecd_002D13e7_002D4dff_002Dbfc9_002De6301d5b749a/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="Adds_Language" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;TestAncestor&gt;
-    &lt;TestId&gt;xUnit::F8BB8A80-6BF0-47C7-8ADC-CCC226216DCB::net8.0::Web.Test.LanguageControllerTest&lt;/TestId&gt;
-  &lt;/TestAncestor&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=94ce6ecd_002D13e7_002D4dff_002Dbfc9_002De6301d5b749a/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="Adds_Language" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;Project Location="C:\Users\domin\RiderProjects\AutoFieldTranslationExperiment\test\Web.Integration" Presentation="&amp;lt;test&amp;gt;\&amp;lt;Web.Integration&amp;gt;" /&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=f478eb2c_002D0fd8_002D4e75_002D9b3f_002D33713a1fb14b/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="AddMethod" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;TestAncestor&gt;
-    &lt;TestId&gt;xUnit::F8BB8A80-6BF0-47C7-8ADC-CCC226216DCB::net8.0::Web.Test.LanguageControllerTest&lt;/TestId&gt;
-  &lt;/TestAncestor&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=f478eb2c_002D0fd8_002D4e75_002D9b3f_002D33713a1fb14b/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="AddMethod" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;TestAncestor&gt;&#xD;
+    &lt;TestId&gt;xUnit::F8BB8A80-6BF0-47C7-8ADC-CCC226216DCB::net8.0::Web.Test.LanguageControllerTest&lt;/TestId&gt;&#xD;
+  &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
